// src/types/generated/server.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  date: { input: any; output: any; }
  numeric: { input: any; output: any; }
  timestamptz: { input: any; output: any; }
  uuid: { input: any; output: any; }
};
// ...
export type ChatMessageFiles = {
  __typename?: 'ChatMessageFiles';
  archivedAt?: Maybe<Scalars['timestamptz']['output']>;
  chatMessageId: Scalars['uuid']['output'];
  createdAt: Scalars['timestamptz']['output'];
  deletedAt?: Maybe<Scalars['timestamptz']['output']>;
  fileName: Scalars['String']['output'];
  filePath: Scalars['String']['output'];
  fileStorageProvider: Scalars['String']['output'];
  id: Scalars['uuid']['output'];
  /** An object relationship */
  message: ChatMessages;
  mimeType: Scalars['String']['output'];
  originalFileName: Scalars['String']['output'];
  size: Scalars['Int']['output'];
  updatedAt: Scalars['timestamptz']['output'];
};

/** aggregated selection of "chat_message_files" */
export type ChatMessageFilesAggregate = {
  __typename?: 'ChatMessageFilesAggregate';
  aggregate?: Maybe<ChatMessageFilesAggregateFields>;
  nodes: Array<ChatMessageFiles>;
};
// ...
export type ChatMessageFilesInsertInput = {
  archivedAt?: InputMaybe<Scalars['timestamptz']['input']>;
  chatMessageId?: InputMaybe<Scalars['uuid']['input']>;
  createdAt?: InputMaybe<Scalars['timestamptz']['input']>;
  deletedAt?: InputMaybe<Scalars['timestamptz']['input']>;
  fileName?: InputMaybe<Scalars['String']['input']>;
  filePath?: InputMaybe<Scalars['String']['input']>;
  fileStorageProvider?: InputMaybe<Scalars['String']['input']>;
  id?: InputMaybe<Scalars['uuid']['input']>;
  message?: InputMaybe<ChatMessagesObjRelInsertInput>;
  mimeType?: InputMaybe<Scalars['String']['input']>;
  originalFileName?: InputMaybe<Scalars['String']['input']>;
  size?: InputMaybe<Scalars['Int']['input']>;
  updatedAt?: InputMaybe<Scalars['timestamptz']['input']>;
};
// ...
export type ChatMessageFilesMutationResponse = {
  __typename?: 'ChatMessageFilesMutationResponse';
  /** number of rows affected by the mutation */
  affectedRows: Scalars['Int']['output'];
  /** data from the rows affected by the mutation */
  returning: Array<ChatMessageFiles>;
};
// ...
export type ChatMessages = {
  __typename?: 'ChatMessages';
  archivedAt?: Maybe<Scalars['timestamptz']['output']>;
  chatThreadId: Scalars['uuid']['output'];
  completionTokens: Scalars['Int']['output'];
  content: Scalars['String']['output'];
  createdAt: Scalars['timestamptz']['output'];
  deletedAt?: Maybe<Scalars['timestamptz']['output']>;
  /** An array relationship */
  files: Array<ChatMessageFiles>;
  /** An aggregate relationship */
  filesAggregate: ChatMessageFilesAggregate;
  frequencyPenalty?: Maybe<Scalars['numeric']['output']>;
  id: Scalars['uuid']['output'];
  model?: Maybe<Scalars['String']['output']>;
  presencePenalty?: Maybe<Scalars['numeric']['output']>;
  promptTokens: Scalars['Int']['output'];
  role: Scalars['String']['output'];
  temperature?: Maybe<Scalars['numeric']['output']>;
  /** An object relationship */
  thread: ChatThreads;
  topP?: Maybe<Scalars['numeric']['output']>;
  totalTokens: Scalars['Int']['output'];
  updatedAt: Scalars['timestamptz']['output'];
  /** An object relationship */
  user?: Maybe<Users>;
  userId?: Maybe<Scalars['uuid']['output']>;
};
// ...
export type ChatMessagesObjRelInsertInput = {
  data: ChatMessagesInsertInput;
  /** upsert condition */
  onConflict?: InputMaybe<ChatMessagesOnConflict>;
};
// ...
export type ChatThreads = {
  __typename?: 'ChatThreads';
  archivedAt?: Maybe<Scalars['timestamptz']['output']>;
  /** An object relationship */
  breadEntry?: Maybe<BreadEntries>;
  breadEntryId?: Maybe<Scalars['uuid']['output']>;
  createdAt: Scalars['timestamptz']['output'];
  deletedAt?: Maybe<Scalars['timestamptz']['output']>;
  id: Scalars['uuid']['output'];
  /** An array relationship */
  messages: Array<ChatMessages>;
  /** An aggregate relationship */
  messagesAggregate: ChatMessagesAggregate;
  title: Scalars['String']['output'];
  updatedAt: Scalars['timestamptz']['output'];
  /** An object relationship */
  user: Users;
  userId: Scalars['uuid']['output'];
};
// ...
export type Users = {
  __typename?: 'Users';
  birthDate?: Maybe<Scalars['date']['output']>;
  /** An array relationship */
  breadEntries: Array<BreadEntries>;
  /** An aggregate relationship */
  breadEntriesAggregate: BreadEntriesAggregate;
  createdAt: Scalars['timestamptz']['output'];
  deletedAt?: Maybe<Scalars['timestamptz']['output']>;
  email: Scalars['String']['output'];
  id: Scalars['uuid']['output'];
  name: Scalars['String']['output'];
  /** An array relationship */
  recipeNotes: Array<RecipeNotes>;
  /** An aggregate relationship */
  recipeNotesAggregate: RecipeNotesAggregate;
  /** An array relationship */
  recipes: Array<Recipes>;
  /** An aggregate relationship */
  recipesAggregate: RecipesAggregate;
  stripeCustomerId: Scalars['String']['output'];
  updatedAt: Scalars['timestamptz']['output'];
};
// ...
export type InsertChatMessageUserMutationVariables = Exact<{
  id: Scalars['uuid']['input'];
  userId: Scalars['uuid']['input'];
  threadId: Scalars['uuid']['input'];
  role: Scalars['String']['input'];
  content: Scalars['String']['input'];
  promptTokens?: InputMaybe<Scalars['Int']['input']>;
  files?: InputMaybe<Array<ChatMessageFilesInsertInput> | ChatMessageFilesInsertInput>;
  isMedical?: InputMaybe<Scalars['Boolean']['input']>;
  isMedicalConfidence?: InputMaybe<Scalars['numeric']['input']>;
}>;


export type InsertChatMessageUserMutation = { __typename?: 'mutation_root', insertChatMessagesOne?: { __typename?: 'ChatMessages', id: any, content: string, role: string, createdAt: any, updatedAt: any } | null, insertChatMessageFiles?: { __typename?: 'ChatMessageFilesMutationResponse', returning: Array<{ __typename?: 'ChatMessageFiles', id: any }> } | null };